updated: matched product type

+++ src/main/java/com/learningmat/ecommerce/module/inventory/InventoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface InventoryRepository extends JpaRepository<Inventory, Long> {
-    Optional<Inventory> findByProductId(int productId);
+    Optional<Inventory> findByProductId(Long productId);
 }
